VaR_API/requirements.txt
alabaster==1.0.0
babel==2.16.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
Flask==3.1.0
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
packaging==24.2
pandas==2.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tomli==2.2.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
